created notebook-menu-switcher from menu-switcher and moved notebook logic into that component.
